Revert "ka"
This reverts commit 3f61609e1543093a1dd9cb0a63425a879ec343d5.

+++ SoundMotion_v6/Note.pde
@@ -46,7 +46,7 @@ class Note {
     y = _y;
   }
   
-  void setXY(float _x, float _y) {
+  void setXY(int _x, int _y) {
     x = _x;
     y = _y;
   }
